requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.19.0
pandas>=1.2.0
scipy>=1.6.0
pyyaml>=5.4.0
tensorboard>=2.7.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
timm>=0.5.0
spikingjelly>=0.0.0.0.4
